--- a/atlasdb-api/build.gradle
+++ b/atlasdb-api/build.gradle
@@ -17,6 +17,7 @@ dependencies {
   implementation 'com.github.ben-manes.caffeine:caffeine'
   implementation 'com.palantir.conjure.java.api:ssl-config'
   implementation 'com.palantir.conjure.java.runtime:conjure-java-jaxrs-client'
+  implementation 'com.palantir.tritium:tritium-ids'
   implementation ('com.palantir.tritium:tritium-registry') {
     exclude (group: 'io.dropwizard.metrics', module: 'metrics-core')
     exclude (group: 'org.hdrhistogram', module: 'HdrHistogram')