--- a/atlasdb-api/src/main/java/com/palantir/atlasdb/persistent/api/PersistentStore.java
+++ b/atlasdb-api/src/main/java/com/palantir/atlasdb/persistent/api/PersistentStore.java
@@ -19,7 +19,7 @@ package com.palantir.atlasdb.persistent.api;
 import com.google.protobuf.ByteString;
 import com.palantir.common.annotations.ImmutablesStyles.PackageVisibleImmutablesStyle;
 import com.palantir.logsafe.exceptions.SafeIllegalArgumentException;
-import com.palantir.util.UniqueIds;
+import com.palantir.tritium.ids.UniqueIds;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -38,7 +38,7 @@ public interface PersistentStore extends AutoCloseable {
         UUID id();
 
         static Handle newHandle() {
-            return ImmutableHandle.builder().id(UniqueIds.uuid()).build();
+            return ImmutableHandle.builder().id(UniqueIds.randomUuidV4()).build();
         }
     }