--- a/atlasdb-client/build.gradle
+++ b/atlasdb-client/build.gradle
@@ -63,6 +63,7 @@ dependencies {
   implementation 'com.palantir.tracing:tracing-api'
   implementation 'com.palantir.tritium:tritium-api'
   implementation 'com.palantir.tritium:tritium-core'
+  implementation 'com.palantir.tritium:tritium-ids'
   implementation 'com.palantir.tritium:tritium-slf4j'
   implementation 'com.squareup:javapoet'
   implementation 'io.dropwizard.metrics:metrics-core'