--- a/atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/TransactionRetryStrategy.java
+++ b/atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/TransactionRetryStrategy.java
@@ -34,6 +34,7 @@ import com.palantir.exception.NotInitializedException;
 import com.palantir.logsafe.SafeArg;
 import com.palantir.logsafe.logger.SafeLogger;
 import com.palantir.logsafe.logger.SafeLoggerFactory;
+import com.palantir.util.UniqueIds;
 import java.util.Random;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
@@ -68,7 +69,7 @@ public final class TransactionRetryStrategy {
 
     @SuppressWarnings("rawtypes") // StopStrategy uses raw Attempt
     public <T, E extends Exception> T runWithRetry(IntPredicate shouldStopRetrying, Retryable<T, E> task) throws E {
-        UUID runId = UUID.randomUUID();
+        UUID runId = UniqueIds.uuid();
         Retryer<T> retryer = RetryerBuilder.<T>newBuilder()
                 .retryIfException(TransactionRetryStrategy::shouldRetry)
                 .withBlockStrategy(blockStrategy)