--- a/changelog/@unreleased/pr-6726.v2.yml
+++ b/changelog/@unreleased/pr-6726.v2.yml
@@ -1,0 +1,5 @@
+type: improvement
+improvement:
+  description: Add UniqueIds to generate UUIDs with fewer scalability bottlenecks
+  links:
+  - https://github.com/palantir/atlasdb/pull/6726