--- a/lock-api-objects/src/main/java/com/palantir/lock/v2/IdentifiedTimeLockRequest.java
+++ b/lock-api-objects/src/main/java/com/palantir/lock/v2/IdentifiedTimeLockRequest.java
@@ -17,6 +17,7 @@ package com.palantir.lock.v2;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.palantir.util.UniqueIds;
 import java.util.UUID;
 import org.immutables.value.Value;
 
@@ -29,6 +30,6 @@ public interface IdentifiedTimeLockRequest {
     UUID getRequestId();
 
     static IdentifiedTimeLockRequest create() {
-        return ImmutableIdentifiedTimeLockRequest.of(UUID.randomUUID());
+        return ImmutableIdentifiedTimeLockRequest.of(UniqueIds.uuid());
     }
 }