--- a/lock-api-objects/src/main/java/com/palantir/lock/v2/IdentifiedTimeLockRequest.java
+++ b/lock-api-objects/src/main/java/com/palantir/lock/v2/IdentifiedTimeLockRequest.java
@@ -17,7 +17,7 @@ package com.palantir.lock.v2;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import com.palantir.util.UniqueIds;
+import com.palantir.tritium.ids.UniqueIds;
 import java.util.UUID;
 import org.immutables.value.Value;
 
@@ -30,6 +30,6 @@ public interface IdentifiedTimeLockRequest {
     UUID getRequestId();
 
     static IdentifiedTimeLockRequest create() {
-        return ImmutableIdentifiedTimeLockRequest.of(UniqueIds.uuid());
+        return ImmutableIdentifiedTimeLockRequest.of(UniqueIds.pseudoRandomUuidV4());
     }
 }