--- a/lock-api-objects/src/main/java/com/palantir/lock/v2/StartTransactionRequestV4.java
+++ b/lock-api-objects/src/main/java/com/palantir/lock/v2/StartTransactionRequestV4.java
@@ -18,6 +18,7 @@ package com.palantir.lock.v2;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.palantir.util.UniqueIds;
 import java.util.UUID;
 import org.immutables.value.Value;
 
@@ -33,7 +34,7 @@ public interface StartTransactionRequestV4 {
 
     static StartTransactionRequestV4 createForRequestor(UUID requestorUuid, int numTransactions) {
         return ImmutableStartTransactionRequestV4.builder()
-                .requestId(UUID.randomUUID())
+                .requestId(UniqueIds.uuid())
                 .requestorId(requestorUuid)
                 .numTransactions(numTransactions)
                 .build();