--- a/lock-api-objects/src/main/java/com/palantir/lock/v2/WaitForLocksRequest.java
+++ b/lock-api-objects/src/main/java/com/palantir/lock/v2/WaitForLocksRequest.java
@@ -20,7 +20,7 @@ import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.palantir.lock.LockDescriptor;
 import com.palantir.logsafe.Preconditions;
 import com.palantir.logsafe.Unsafe;
-import com.palantir.util.UniqueIds;
+import com.palantir.tritium.ids.UniqueIds;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
@@ -46,13 +46,17 @@ public interface WaitForLocksRequest {
     Optional<String> getClientDescription();
 
     static WaitForLocksRequest of(Set<LockDescriptor> lockDescriptors, long acquireTimeoutMs) {
-        return ImmutableWaitForLocksRequest.of(UniqueIds.uuid(), lockDescriptors, acquireTimeoutMs, Optional.empty());
+        return ImmutableWaitForLocksRequest.of(
+                UniqueIds.pseudoRandomUuidV4(), lockDescriptors, acquireTimeoutMs, Optional.empty());
     }
 
     static WaitForLocksRequest of(
             Set<LockDescriptor> lockDescriptors, long acquireTimeoutMs, @Nullable String clientDescription) {
         return ImmutableWaitForLocksRequest.of(
-                UniqueIds.uuid(), lockDescriptors, acquireTimeoutMs, Optional.ofNullable(clientDescription));
+                UniqueIds.pseudoRandomUuidV4(),
+                lockDescriptors,
+                acquireTimeoutMs,
+                Optional.ofNullable(clientDescription));
     }
 
     @Value.Check