--- a/lock-api/src/main/java/com/palantir/lock/client/IdentifiedLockRequest.java
+++ b/lock-api/src/main/java/com/palantir/lock/client/IdentifiedLockRequest.java
@@ -22,6 +22,7 @@ import com.palantir.lock.LockDescriptor;
 import com.palantir.lock.v2.LockRequest;
 import com.palantir.lock.watch.LockRequestMetadata;
 import com.palantir.logsafe.Unsafe;
+import com.palantir.util.UniqueIds;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
@@ -51,13 +52,13 @@ public interface IdentifiedLockRequest {
 
     static IdentifiedLockRequest of(Set<LockDescriptor> lockDescriptors, long acquireTimeoutMs) {
         return ImmutableIdentifiedLockRequest.of(
-                UUID.randomUUID(), lockDescriptors, acquireTimeoutMs, Optional.empty(), Optional.empty());
+                UniqueIds.uuid(), lockDescriptors, acquireTimeoutMs, Optional.empty(), Optional.empty());
     }
 
     static IdentifiedLockRequest of(
             Set<LockDescriptor> lockDescriptors, long acquireTimeoutMs, String clientDescription) {
         return ImmutableIdentifiedLockRequest.of(
-                UUID.randomUUID(), lockDescriptors, acquireTimeoutMs, Optional.of(clientDescription), Optional.empty());
+                UniqueIds.uuid(), lockDescriptors, acquireTimeoutMs, Optional.of(clientDescription), Optional.empty());
     }
 
     static IdentifiedLockRequest of(
@@ -66,7 +67,7 @@ public interface IdentifiedLockRequest {
             String clientDescription,
             LockRequestMetadata metadata) {
         return ImmutableIdentifiedLockRequest.of(
-                UUID.randomUUID(),
+                UniqueIds.uuid(),
                 lockDescriptors,
                 acquireTimeoutMs,
                 Optional.of(clientDescription),