--- a/lock-api/src/main/java/com/palantir/lock/client/LockLeaseService.java
+++ b/lock-api/src/main/java/com/palantir/lock/client/LockLeaseService.java
@@ -37,7 +37,7 @@ import com.palantir.lock.v2.WaitForLocksRequest;
 import com.palantir.lock.v2.WaitForLocksResponse;
 import com.palantir.lock.watch.LockWatchVersion;
 import com.palantir.logsafe.Preconditions;
-import com.palantir.util.UniqueIds;
+import com.palantir.tritium.ids.UniqueIds;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
@@ -68,7 +68,7 @@ public class LockLeaseService implements AutoCloseable {
             NamespacedConjureTimelockService conjureTimelock,
             LeaderTimeGetter leaderTimeGetter,
             LockTokenUnlocker unlocker) {
-        return new LockLeaseService(conjureTimelock, UniqueIds.uuid(), leaderTimeGetter, unlocker);
+        return new LockLeaseService(conjureTimelock, UniqueIds.pseudoRandomUuidV4(), leaderTimeGetter, unlocker);
     }
 
     LockImmutableTimestampResponse lockImmutableTimestamp() {
@@ -78,7 +78,7 @@ public class LockLeaseService implements AutoCloseable {
     StartTransactionResponseV4 startTransactions(int batchSize) {
         ConjureStartTransactionsRequest request = ConjureStartTransactionsRequest.builder()
                 .requestorId(clientId)
-                .requestId(UniqueIds.uuid())
+                .requestId(UniqueIds.pseudoRandomUuidV4())
                 .numTransactions(batchSize)
                 .lastKnownVersion(Optional.empty())
                 .build();
@@ -100,7 +100,7 @@ public class LockLeaseService implements AutoCloseable {
             Optional<LockWatchVersion> maybeVersion, int batchSize) {
         ConjureStartTransactionsRequest request = ConjureStartTransactionsRequest.builder()
                 .requestorId(clientId)
-                .requestId(UniqueIds.uuid())
+                .requestId(UniqueIds.pseudoRandomUuidV4())
                 .numTransactions(batchSize)
                 .lastKnownVersion(ConjureLockRequests.toConjure(maybeVersion))
                 .build();