@schlosna 2023-09-06
Pending https://github.com/palantir/tritium/pull/1789